fix: save main functions to __fish_cd.fish and __plugin_cd.fish
Fish automatically searches through any directories in
$fish_function_path, and any functions defined are
automatically loaded when needed. A function definition
file must have a filename consisting of the name of the
function plus the suffix '.fish'.

It is very important that function definition files contain
the definition for the specified function and nothing else.
Otherwise, it is possible that autoloading a function files
requires that the function already be loaded, which creates
a circular dependency.